Reorganise stuff
